build: enable bevy dynamic linking
Enabling this feature requires the nightly toolchain, so a
rust-toolchain file has been added to the repo to ensure the override is
in place for the project.

This diff enables the feature by default, but this might break building
for wasm. If it does, we can (at time of writing) pass
`--no-default-features`, or otherwise we'll just have to change the
feature list in the manifest.

For now, it should improve build times while we iterate.