feat: initial attempt at fatigue tracking
Finding the player "pushing" state to not be as reliable for
manipulating a fatigue value as I'd like. Not sure what the story is.

It might be better to leverage rapier events [1], but we may also need
to define some collider groups so as to limit the events to just those
between the player and the boulder. TBD what the right approach is.

OTOH, the tick rates of the system might be the reason why the rate of
change seems unreliable. The reasons are still unclear.

[1] https://rapier.rs/docs/user_guides/bevy_plugin/advanced_collision_detection#collision-and-contact-force-events

---

Edit: looks like we can somewhat get away with simply moving the
update_fatigue system out of the `FixedUpdate` schedule and into
`Update` so it samples the `PlayerState` a lot more often.

Still not clear on what to do when the fatigue limit hits, but I'll work
that out in a follow-up.