feat: knock-back when fatigue approaches max
Adds an `ExternalForce` component to the player, initialized with a
torque of zero. When we hit the fatigue limit, we raise the torque to
apply force to the player (the appearance is sort of like rolling
backward downhill). The torque must be reset to zero promptly to prevent
perpetual spinning.

A `Damping` comonent is also added to limit the overall motion of the
force when applied.